rename: style 확장자를 styled로 변경
