+++ frontend/src/components/DiscussionSubmit/DiscussionTitle.tsx
@@ -1,7 +1,7 @@
 import { ERROR_MESSAGE } from '@/constants/messages';
 import type { ChangeEventHandler } from 'react';
 import Input from '@/components/common/Input/Input';
-import * as S from './DiscussionSubmit.style';
+import * as S from './DiscussionSubmit.styled';
 
 interface DiscussionTitleProps {
   value: string;
@@ -17,7 +17,7 @@ export default function DiscussionTitle({ danger, value, onChange }: DiscussionT
         width="xlarge"
         danger={danger}
         dangerMessage={ERROR_MESSAGE.invalid_title}
-        placeholder='글 제목을 입력해주세요'
+        placeholder="글 제목을 입력해주세요"
         value={value}
         onChange={onChange}
       />
