+++ frontend/src/components/DiscussionSubmit/index.tsx
@@ -8,7 +8,7 @@ import useMissions from '@/hooks/useMissions';
 import TagMultipleList from '../common/TagMultipleList';
 import type { HashTag } from '@/types';
 import TagList from '@/components/common/TagList';
-import * as S from './DiscussionSubmit.style';
+import * as S from './DiscussionSubmit.styled';
 import { useSearchParams } from 'react-router-dom';
 import useDiscussion from '@/hooks/useDiscussion';
 import useUserInfo from '@/hooks/useUserInfo';
@@ -98,12 +98,14 @@ export default function DiscussionSubmit() {
           selectedTag={selectedMission}
           setSelectedTag={setSelectedMission}
           variant="danger"
+          label="미션"
           keyName="title"
         />
         <TagMultipleList
           tags={allHashTags}
           selectedTags={selectedHashTags}
           setSelectedTags={setSelectedHashTags}
+          label="해시 태그"
           keyName="name"
         />
       </S.DiscussionTagListWrapper>
